+++ lib/features/admin/screens/teacher/teacher_management_screen.dart
@@ -2,6 +2,7 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 
 import '../../../../core/services/teacher_service.dart';
+import '../../../../shared/widgets/delete_confirmation_dialog.dart';
 
 class TeacherManagementScreen extends StatefulWidget {
   const TeacherManagementScreen({super.key});
@@ -59,6 +60,10 @@ class _TeacherManagementScreenState extends State<TeacherManagementScreen> {
   }
 
   Future<void> deleteTeacher(String teacherId) async {
+    final confirmed = await showDeleteConfirmationDialog(context, "Teacher");
+
+    if (confirmed != true) return;
+
     await _teacherService.deleteTeacher(teacherId);
 
     if (!mounted) return;
